requirements.txt
streamlit
groq
python-dotenv
faiss-cpu
sentence-transformers
paddleocr
paddlepaddle
openai-whisper
pydub
sqlite-utils